Fixed some UI and file management
Adjusted labels to fit large names
Fixed recordscene to look and work better
moved all folders in to a Resources folder
added some more alerts
fixed a few bugs with upload controller
added loading indicators in upload/playlist

+++ AnotherDuplicatelol/src/namesayer/fxmlControllers/DeletePlaylistController.java
@@ -6,6 +6,7 @@ import java.io.PrintWriter;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -16,7 +17,9 @@ import javafx.fxml.Initializable;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
 import javafx.stage.Stage;
 import namesayer.helperClasses.DatabaseList;
@@ -44,21 +47,29 @@ public class DeletePlaylistController implements Initializable{
 
     @FXML
     public void replaceButtonClicked(ActionEvent event) {
-        String nameToDelete = playlistView.getSelectionModel().getSelectedItem();
-        deletePlaylist(nameToDelete + ".txt");
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("fxmlFiles/SaveScreen.fxml"));
-            Parent rateSceneParent = fxmlLoader.load();
-            SaveScreenController controller = fxmlLoader.getController();
-            controller.setRecordingList(listOfRecordings);
-            controller.setButton(workspaceButton);
-            Scene rateScene = new Scene(rateSceneParent);
-            Stage rateStage = (Stage) replaceButton.getScene().getWindow();;
-            rateStage.setScene(rateScene);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Delete Playlist");
+        alert.setHeaderText("Continuing will delete this playlist");
+        alert.setContentText("Are you sure you want to continue?");
+        Optional<ButtonType> action = alert.showAndWait();
 
+        if (action.get() == ButtonType.OK) {
+            String nameToDelete = playlistView.getSelectionModel().getSelectedItem();
+            deletePlaylist(nameToDelete + ".txt");
+            try {
+                FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("fxmlFiles/SaveScreen.fxml"));
+                Parent rateSceneParent = fxmlLoader.load();
+                SaveScreenController controller = fxmlLoader.getController();
+                controller.setRecordingList(listOfRecordings);
+                controller.setButton(workspaceButton);
+                Scene rateScene = new Scene(rateSceneParent);
+                Stage rateStage = (Stage) replaceButton.getScene().getWindow();
+                ;
+                rateStage.setScene(rateScene);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
     @FXML
@@ -80,7 +91,7 @@ public class DeletePlaylistController implements Initializable{
     }
 
     public void createPlaylist(String name) {
-        File file = new File("./Playlists/" + name + ".txt");
+        File file = new File("./Resources/Playlists/" + name + ".txt");
         try {
             PrintWriter writer = new PrintWriter(file);
             for (String currentName : listOfRecordings.getRecordingNames()) {
@@ -93,7 +104,7 @@ public class DeletePlaylistController implements Initializable{
     }
 
     public void deletePlaylist(String name) {
-        File playlistFiles = new File("Playlists/");
+        File playlistFiles = new File("Resources/Playlists/");
         for (File fileEntry : playlistFiles.listFiles()) {
             String currentFile = fileEntry.getName();
             if (currentFile.equals(name)) {
@@ -103,7 +114,7 @@ public class DeletePlaylistController implements Initializable{
     }
 
     public void listPlaylists() {
-        File playlistFiles = new File("Playlists/");
+        File playlistFiles = new File("Resources/Playlists/");
         for (final File fileEntry : playlistFiles.listFiles()) {
             String currentFile = fileEntry.getName();
             if (currentFile.endsWith(".txt")) {
